This is the implementation of the network based database, and fallback sqlite if unable to connect to database.
Login works from the asp.net half and on correct login sends you back to asp.net home with a message of who you are logged in as.
On error it says there was an error.
TODO: use cookies to stay logged in with a sessionid
TODO: populate the page based on logged in account

Local app is able to create logins, and displays before and afters of the logins table

Note: After this we should start making branches and merging.