Added Log out functionallity. Added Change password functionallity to the account page.
Added display of username in the layout.cshtml if you are logged in.

Display login or logout depending on login status.

Added checkPasswordDB() to databasecreator.cs for checking password. TODO make this check hash instead possibly.

Added updatePasswordDB() to databasecreator.cs Again TODO change this to hash possibly.

Require log in to be able to view quizzes/account page.