Added methods to database creator to handle grading and submitting quizzes ++AddGradeDB() adds a row to grade table with specified values ++AddResponseDB() adds rows to response table with specified values for each response in the list ++gradeQuizDB() grades the specified quiz_id against the list of response. Then calls addGradeDB() and AddResponseDB() to add the grades and responses to the table.
Have the controller action for submit quiz call gradeQuizDB() then redirect the user to a page showing the grade afterwards.

Rewrote parts of the homecontroller to change data types of a few variables from string[,] to list<string[]> for easier access and to avoid null array rows

Added question count to quiz page, correctly set values of form items in the question type_ID partial views to be able to get the normal response values in the controller instead of the randomized values that are displayed.